requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
anyio==2.1.0              # via jupyter-server
appdirs==1.4.4            # via black
appnope==0.1.2            # via ipykernel, ipython
argon2-cffi==20.1.0       # via notebook
async-generator==1.10     # via nbclient
attrs==20.3.0             # via jsonschema
babel==2.9.0              # via jupyterlab-server
backcall==0.2.0           # via ipython
black==20.8b1             # via -r requirements.in, nb-black
bleach==3.3.0             # via nbconvert
certifi==2020.12.5        # via requests
cffi==1.14.5              # via argon2-cffi
chardet==4.0.0            # via requests
click==7.1.2              # via black
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
fsspec==0.8.5             # via -r requirements.in
gin-config==0.4.0         # via -r requirements.in
idna==2.10                # via anyio, requests
importlib-metadata==3.4.0  # via jsonschema
ipykernel==5.4.3          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via jupyter-server, nbformat, notebook, qtconsole, traitlets
ipython==7.20.0           # via ipykernel, ipywidgets, jupyter-console, jupyterlab, nb-black
ipywidgets==7.6.3         # via jupyter
jedi==0.18.0              # via ipython
jinja2==2.11.3            # via jupyter-server, jupyterlab, jupyterlab-server, nbconvert, notebook
json5==0.9.5              # via jupyterlab-server
jsonschema==3.2.0         # via jupyterlab-server, nbformat
jupyter-client==6.1.11    # via ipykernel, jupyter-console, jupyter-server, nbclient, notebook, qtconsole
jupyter-console==6.2.0    # via jupyter
jupyter-core==4.7.1       # via jupyter-client, jupyter-server, jupyterlab, nbconvert, nbformat, notebook, qtconsole
jupyter-server==1.3.0     # via jupyterlab, jupyterlab-server, nbclassic
jupyter==1.0.0            # via -r requirements.in
jupyterlab-pygments==0.1.2  # via nbconvert
jupyterlab-server==2.2.0  # via jupyterlab
jupyterlab-widgets==1.0.0  # via ipywidgets
jupyterlab==3.0.7         # via -r requirements.in
markupsafe==1.1.1         # via jinja2
mistune==0.8.4            # via nbconvert
mypy-extensions==0.4.3    # via black
nb-black==1.0.7           # via -r requirements.in
nbclassic==0.2.6          # via jupyterlab
nbclient==0.5.2           # via nbconvert
nbconvert==6.0.7          # via jupyter, jupyter-server, notebook
nbformat==5.1.2           # via ipywidgets, jupyter-server, nbclient, nbconvert, notebook
nest-asyncio==1.5.1       # via nbclient
notebook==6.2.0           # via jupyter, nbclassic, widgetsnbextension
numpy==1.20.1             # via -r requirements.in, pandas, patsy, scipy, statsmodels
packaging==20.9           # via bleach, jupyterlab, jupyterlab-server
pandas==1.2.2             # via -r requirements.in, statsmodels
pandocfilters==1.4.3      # via nbconvert
parso==0.8.1              # via jedi
pathspec==0.8.1           # via black
patsy==0.5.1              # via statsmodels
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
plotly==4.14.3            # via -r requirements.in
prometheus-client==0.9.0  # via jupyter-server, notebook
prompt-toolkit==3.0.16    # via ipython, jupyter-console
ptyprocess==0.7.0         # via pexpect, terminado
pycparser==2.20           # via cffi
pygments==2.7.4           # via ipython, jupyter-console, jupyterlab-pygments, nbconvert, qtconsole
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
python-dateutil==2.8.1    # via jupyter-client, pandas
pytz==2021.1              # via babel, pandas
pyzmq==22.0.3             # via jupyter-client, jupyter-server, notebook, qtconsole
qtconsole==5.0.2          # via jupyter
qtpy==1.9.0               # via qtconsole
regex==2020.11.13         # via black
requests==2.25.1          # via jupyterlab-server
retrying==1.3.3           # via plotly
scipy==1.6.0              # via -r requirements.in, statsmodels
send2trash==1.5.0         # via jupyter-server, notebook
six==1.15.0               # via argon2-cffi, bleach, jsonschema, patsy, plotly, python-dateutil, retrying
sniffio==1.2.0            # via anyio
statsmodels==0.12.2       # via -r requirements.in
tabulate==0.8.7           # via -r requirements.in
terminado==0.9.2          # via jupyter-server, notebook
testpath==0.4.4           # via nbconvert
toml==0.10.2              # via black
tornado==6.1              # via ipykernel, jupyter-client, jupyter-server, jupyterlab, notebook, terminado
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, jupyter-server, nbclient, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.4.2          # via black
typing-extensions==3.7.4.3  # via anyio, black, importlib-metadata
urllib3==1.26.3           # via requests
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
widgetsnbextension==3.5.1  # via ipywidgets
zipp==3.4.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
